Move scheduler constants into their own module so that packages with native dependencies are not transitively included in the package start-up script. (#80)
