+++ components/job-orchestration/job_orchestration/scheduler/scheduler_data.py
@@ -8,26 +8,11 @@ import zstandard
 from celery.result import AsyncResult
 from pydantic import BaseModel, validator
 
-TASK_QUEUE_LOWEST_PRIORITY = 1
-TASK_QUEUE_HIGHEST_PRIORITY = 3
-
-
-class QueueName:
-    COMPRESSION = "compression"
-    SEARCH = "search"
-
-
-class TaskUpdateType:
-    COMPRESSION = 'COMPRESSION'
-    SEARCH = 'SEARCH'
-
-
-class TaskStatus:
-    SUBMITTED = 'SUBMITTED'
-    SCHEDULED = 'SCHEDULED'
-    IN_PROGRESS = 'IN_PROGRESS'
-    SUCCEEDED = 'SUCCEEDED'
-    FAILED = 'FAILED'
+from .constants import \
+    TASK_QUEUE_LOWEST_PRIORITY, \
+    TASK_QUEUE_HIGHEST_PRIORITY, \
+    TaskStatus, \
+    TaskUpdateType
 
 
 class TaskUpdate(BaseModel):
@@ -91,13 +76,6 @@ class SearchTask(BaseModel):
     # pydantic has no validator for it
     class Config:
         arbitrary_types_allowed = True
-
-
-class JobStatus:
-    SCHEDULING = 'SCHEDULING'
-    SCHEDULED = 'SCHEDULED'
-    SUCCEEDED = 'SUCCEEDED'
-    FAILED = 'FAILED'
 
 
 class CompressionJob(BaseModel):
